+++ test/Sentry.Tests/ScopeTests.cs
@@ -345,34 +345,6 @@ public class ScopeTests
     [Theory]
     [InlineData(true)]
     [InlineData(false)]
-    public void ResetTransaction_MatchingTransaction_ObserverSetsTraceFromPropagationContextIfEnabled(bool enableScopeSync)
-    {
-        // Arrange
-        var observer = Substitute.For<IScopeObserver>();
-        var scope = new Scope(new SentryOptions
-        {
-            ScopeObserver = observer,
-            EnableScopeSync = enableScopeSync
-        });
-        var transaction = new TransactionTracer(DisabledHub.Instance, "test-transaction", "op");
-        scope.Transaction = transaction;
-
-        var expectedTraceId = scope.PropagationContext.TraceId;
-        var expectedSpanId = scope.PropagationContext.SpanId;
-        var expectedCount = enableScopeSync ? 1 : 0;
-
-        observer.ClearReceivedCalls();
-
-        // Act
-        scope.ResetTransaction(transaction);
-
-        // Assert
-        observer.Received(expectedCount).SetTrace(Arg.Is(expectedTraceId), Arg.Is(expectedSpanId));
-    }
-
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
     public void ResetTransaction_NonMatchingTransaction_ObserverNotCalled(bool enableScopeSync)
     {
         // Arrange